Move a subset of Gradle API out of `core-api` into `core-api-java`
