+++ subprojects/core-api-java/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id("gradlebuild.distribution.api-java")
+}
+
+description = "Kotlin extensions to make working with Gradle :core and family more convenient"
+
+dependencies {
+    implementation(projects.baseServices)
+    compileOnly(libs.jetbrainsAnnotations)
+}
