+++ subprojects/core-api-java/src/main/java/org/gradle/api/provider/Property.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,8 +17,6 @@
 package org.gradle.api.provider;
 
 import org.gradle.api.SupportsKotlinAssignmentOverloading;
-import org.gradle.api.model.ObjectFactory;
-import org.gradle.internal.extensions.stdlib.Assignable;
 
 import javax.annotation.Nullable;
 
@@ -48,7 +46,7 @@ import javax.annotation.Nullable;
  * @since 4.3
  */
 @SupportsKotlinAssignmentOverloading
-public interface Property<T> extends Provider<T>, HasConfigurableValue, SupportsConvention, Assignable<T, Provider<? extends T>> {
+public interface Property<T> extends Provider<T>, HasConfigurableValue, SupportsConvention {
     /**
      * Sets the value of the property to the given value, replacing whatever value the property already had.
      *
@@ -60,7 +58,6 @@ public interface Property<T> extends Provider<T>, HasConfigurableValue, Supports
      *
      * @param value The value, can be null.
      */
-    @Override
     void set(@Nullable T value);
 
     /**
@@ -81,7 +78,6 @@ public interface Property<T> extends Provider<T>, HasConfigurableValue, Supports
      *
      * @param provider The provider of the property's value, can't be null.
      */
-    @Override
     void set(Provider<? extends T> provider);
 
     /**
