+++ subprojects/core-api/build.gradle.kts
@@ -18,6 +18,7 @@ errorprone {
 dependencies {
     compileOnly(libs.jetbrainsAnnotations)
 
+    api(projects.coreApiJava)
     api(projects.stdlibJavaExtensions)
     api(projects.stdlibKotlinExtensions)
     api(projects.buildCacheSpi)
